+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_responses.hpp
@@ -16,16 +16,15 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
     // ServiceClient models:
 
-    using ListSharesSinglePageResult = _detail::ServiceListSharesSinglePageResult;
-    using SetServicePropertiesResult = _detail::ServiceSetPropertiesResult;
-    using GetServicePropertiesResult = FileServiceProperties;
+    using ListSharesSinglePageResult = Details::ServiceListSharesSinglePageResult;
+    using SetServicePropertiesResult = Details::ServiceSetPropertiesResult;
 
     // ShareClient models:
     struct CreateShareResult
     {
       bool Created = true;
       Azure::ETag ETag;
-      DateTime LastModified;
+      Core::DateTime LastModified;
       std::string RequestId;
     };
 
@@ -34,27 +33,25 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
       bool Deleted = true;
       std::string RequestId;
     };
-    using CreateShareSnapshotResult = _detail::ShareCreateSnapshotResult;
-    using GetSharePropertiesResult = _detail::ShareGetPropertiesResult;
-    using SetSharePropertiesResult = _detail::ShareSetPropertiesResult;
-    using SetShareMetadataResult = _detail::ShareSetMetadataResult;
-    using SetShareAccessPolicyResult = _detail::ShareSetAccessPolicyResult;
-    using GetShareStatisticsResult = _detail::ShareGetStatisticsResult;
-    using CreateSharePermissionResult = _detail::ShareCreatePermissionResult;
-    using GetShareAccessPolicyResult = _detail::ShareGetAccessPolicyResult;
-    using GetSharePermissionResult = _detail::ShareGetPermissionResult;
-    using AcquireShareLeaseResult = _detail::ShareAcquireLeaseResult;
-    using RenewShareLeaseResult = _detail::ShareRenewLeaseResult;
-    using ReleaseShareLeaseResult = _detail::ShareReleaseLeaseResult;
-    using BreakShareLeaseResult = _detail::ShareBreakLeaseResult;
-    using ChangeShareLeaseResult = _detail::ShareChangeLeaseResult;
+    using CreateShareSnapshotResult = Details::ShareCreateSnapshotResult;
+    using ShareProperties = Details::ShareGetPropertiesResult;
+    using SetSharePropertiesResult = Details::ShareSetPropertiesResult;
+    using SetShareMetadataResult = Details::ShareSetMetadataResult;
+    using SetShareAccessPolicyResult = Details::ShareSetAccessPolicyResult;
+    using ShareStatistics = Details::ShareGetStatisticsResult;
+    using CreateSharePermissionResult = Details::ShareCreatePermissionResult;
+    using AcquireShareLeaseResult = Details::ShareAcquireLeaseResult;
+    using RenewShareLeaseResult = Details::ShareRenewLeaseResult;
+    using ReleaseShareLeaseResult = Details::ShareReleaseLeaseResult;
+    using BreakShareLeaseResult = Details::ShareBreakLeaseResult;
+    using ChangeShareLeaseResult = Details::ShareChangeLeaseResult;
 
     // DirectoryClient models:
 
     struct CreateShareDirectoryResult
     {
       Azure::ETag ETag;
-      DateTime LastModified;
+      Core::DateTime LastModified;
       std::string RequestId;
       bool IsServerEncrypted = bool();
       FileSmbProperties SmbProperties;
@@ -68,11 +65,11 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
       std::string RequestId;
     };
 
-    using GetShareDirectoryPropertiesResult = _detail::DirectoryGetPropertiesResult;
-    using SetShareDirectoryPropertiesResult = _detail::DirectorySetPropertiesResult;
-    using SetShareDirectoryMetadataResult = _detail::DirectorySetMetadataResult;
+    using ShareDirectoryProperties = Details::DirectoryGetPropertiesResult;
+    using SetShareDirectoryPropertiesResult = Details::DirectorySetPropertiesResult;
+    using SetShareDirectoryMetadataResult = Details::DirectorySetMetadataResult;
     using ForceCloseAllShareDirectoryHandlesSinglePageResult
-        = _detail::DirectoryForceCloseHandlesResult;
+        = Details::DirectoryForceCloseHandlesResult;
 
     struct ForceCloseShareDirectoryHandleResult
     {
@@ -87,7 +84,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
       std::string DirectoryPath;
       std::string Prefix;
       int32_t PageSizeHint = int32_t();
-      Azure::Nullable<std::string> ContinuationToken;
+      Azure::Core::Nullable<std::string> ContinuationToken;
       std::vector<DirectoryItem> DirectoryItems;
       std::vector<FileItem> FileItems;
       std::string RequestId;
@@ -96,7 +93,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     struct ListShareDirectoryHandlesSinglePageResult
     {
       std::vector<HandleItem> Handles;
-      Azure::Nullable<std::string> ContinuationToken;
+      Azure::Core::Nullable<std::string> ContinuationToken;
       std::string RequestId;
     };
 
@@ -105,7 +102,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     {
       bool Created = true;
       Azure::ETag ETag;
-      DateTime LastModified;
+      Core::DateTime LastModified;
       bool IsServerEncrypted = bool();
       FileSmbProperties SmbProperties;
       std::string RequestId;
@@ -119,56 +116,56 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
     struct DownloadShareFileDetails
     {
-      DateTime LastModified;
+      Core::DateTime LastModified;
       Storage::Metadata Metadata;
       Azure::ETag ETag;
-      Azure::Nullable<DateTime> CopyCompletedOn;
-      Azure::Nullable<std::string> CopyStatusDescription;
-      Azure::Nullable<std::string> CopyId;
-      Azure::Nullable<std::string> CopyProgress;
-      Azure::Nullable<std::string> CopySource;
-      Azure::Nullable<CopyStatusType> CopyStatus;
+      Azure::Core::Nullable<Core::DateTime> CopyCompletedOn;
+      Azure::Core::Nullable<std::string> CopyStatusDescription;
+      Azure::Core::Nullable<std::string> CopyId;
+      Azure::Core::Nullable<std::string> CopyProgress;
+      Azure::Core::Nullable<std::string> CopySource;
+      Azure::Core::Nullable<CopyStatusType> CopyStatus;
       bool IsServerEncrypted = bool();
       FileSmbProperties SmbProperties;
-      Azure::Nullable<LeaseDurationType> LeaseDuration;
-      Azure::Nullable<LeaseStateType> LeaseState;
-      Azure::Nullable<LeaseStatusType> LeaseStatus;
+      Azure::Core::Nullable<LeaseDurationType> LeaseDuration;
+      Azure::Core::Nullable<LeaseStateType> LeaseState;
+      Azure::Core::Nullable<LeaseStatusType> LeaseStatus;
     };
 
     struct DownloadShareFileResult
     {
-      std::unique_ptr<Azure::Core::IO::BodyStream> BodyStream;
-      Azure::Core::Http::HttpRange ContentRange;
+      std::unique_ptr<Azure::IO::BodyStream> BodyStream;
+      Azure::Core::Http::Range ContentRange;
       int64_t FileSize = 0;
-      Azure::Nullable<Storage::ContentHash> TransactionalContentHash;
+      Azure::Core::Nullable<Storage::ContentHash> TransactionalContentHash;
       FileHttpHeaders HttpHeaders;
       DownloadShareFileDetails Details;
       std::string RequestId;
     };
 
-    using StartCopyShareFileResult = _detail::FileStartCopyResult;
-    using AbortCopyShareFileResult = _detail::FileAbortCopyResult;
-    using GetShareFilePropertiesResult = _detail::FileGetPropertiesResult;
-    using SetShareFilePropertiesResult = _detail::FileSetHttpHeadersResult;
-    using ResizeFileResult = _detail::FileSetHttpHeadersResult;
-    using SetShareFileMetadataResult = _detail::FileSetMetadataResult;
-    using UploadShareFileRangeResult = _detail::FileUploadRangeResult;
+    using StartCopyShareFileResult = Details::FileStartCopyResult;
+    using AbortCopyShareFileResult = Details::FileAbortCopyResult;
+    using ShareFileProperties = Details::FileGetPropertiesResult;
+    using SetShareFilePropertiesResult = Details::FileSetHttpHeadersResult;
+    using ResizeFileResult = Details::FileSetHttpHeadersResult;
+    using SetShareFileMetadataResult = Details::FileSetMetadataResult;
+    using UploadShareFileRangeResult = Details::FileUploadRangeResult;
     struct ClearShareFileRangeResult
     {
       Azure::ETag ETag;
-      DateTime LastModified;
+      Core::DateTime LastModified;
       std::string RequestId;
       bool IsServerEncrypted = bool();
     };
-    using UploadFileRangeFromUriResult = _detail::FileUploadRangeFromUrlResult;
-    using GetShareFileRangeListResult = _detail::FileGetRangeListResult;
+    using UploadFileRangeFromUriResult = Details::FileUploadRangeFromUrlResult;
+    using ShareFileRangeInfo = Details::FileGetRangeListResult;
     using ListShareFileHandlesSinglePageResult = ListShareDirectoryHandlesSinglePageResult;
-    using ForceCloseAllShareFileHandlesSinglePageResult = _detail::FileForceCloseHandlesResult;
+    using ForceCloseAllShareFileHandlesSinglePageResult = Details::FileForceCloseHandlesResult;
 
     struct DownloadShareFileToResult
     {
       int64_t FileSize = 0;
-      Azure::Core::Http::HttpRange ContentRange;
+      Azure::Core::Http::Range ContentRange;
       FileHttpHeaders HttpHeaders;
       DownloadShareFileDetails Details;
     };
@@ -185,18 +182,24 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
   } // namespace Models
 
-  class StartCopyShareFileOperation
-      : public Azure::Core::Operation<Models::GetShareFilePropertiesResult> {
+  class StartCopyShareFileOperation : public Azure::Core::Operation<Models::ShareFileProperties> {
   public:
     std::string RequestId;
     Azure::ETag ETag;
-    Azure::DateTime LastModified;
+    Azure::Core::DateTime LastModified;
     std::string CopyId;
     Models::CopyStatusType CopyStatus;
-    Azure::Nullable<std::string> VersionId;
+    Azure::Core::Nullable<std::string> VersionId;
 
   public:
-    Models::GetShareFilePropertiesResult Value() const override { return m_pollResult; }
+    Models::ShareFileProperties Value() const override { return m_pollResult; }
+
+    /**
+     * @brief Get the raw HTTP response.
+     * @return A pointer to #Azure::Core::Http::RawResponse.
+     * @note Does not give up ownership of the RawResponse.
+     */
+    Azure::Core::Http::RawResponse* GetRawResponse() const override { return m_rawResponse.get(); }
 
     StartCopyShareFileOperation() = default;
 
@@ -216,22 +219,13 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     std::unique_ptr<Azure::Core::Http::RawResponse> PollInternal(
         Azure::Core::Context& context) override;
 
-    Azure::Response<Models::GetShareFilePropertiesResult> PollUntilDoneInternal(
+    Azure::Response<Models::ShareFileProperties> PollUntilDoneInternal(
         std::chrono::milliseconds period,
         Azure::Core::Context& context) override;
 
-    /**
-     * @brief Get the raw HTTP response.
-     * @return A reference to an #Azure::Core::Http::RawResponse.
-     * @note Does not give up ownership of the RawResponse.
-     */
-    Azure::Core::Http::RawResponse const& GetRawResponseInternal() const override
-    {
-      return *m_rawResponse;
-    }
-
+    std::unique_ptr<Azure::Core::Http::RawResponse> m_rawResponse;
     std::shared_ptr<ShareFileClient> m_fileClient;
-    Models::GetShareFilePropertiesResult m_pollResult;
+    Models::ShareFileProperties m_pollResult;
 
     friend class ShareFileClient;
   };