+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_service_client.hpp
@@ -95,10 +95,9 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief Get the service's properties.
      * @param options Optional parameters to get the properties of the service.
      * @param context Context for cancelling long running operations.
-     * @return Azure::Response<Models::GetServicePropertiesResult> The result containing
-     * service's properties.
+     * @return Azure::Response<Models::FileServiceProperties> The properties of the service.
      */
-    Azure::Response<Models::GetServicePropertiesResult> GetProperties(
+    Azure::Response<Models::FileServiceProperties> GetProperties(
         const GetServicePropertiesOptions& options = GetServicePropertiesOptions(),
         const Azure::Core::Context& context = Azure::Core::Context()) const;