+++ sdk/storage/azure-storage-files-shares/test/share_client_test.cpp
@@ -206,15 +206,11 @@ namespace Azure { namespace Storage { namespace Test {
       identifiers.emplace_back(identifier);
     }
 
-    auto lmt = m_shareClient->GetAccessPolicy()->LastModified;
     auto ret = m_shareClient->SetAccessPolicy(identifiers);
     EXPECT_TRUE(ret->ETag.HasValue());
-    EXPECT_FALSE(ret->LastModified < lmt);
 
-    auto ret2 = m_shareClient->GetAccessPolicy();
-    EXPECT_EQ(ret2->ETag, ret->ETag);
-    EXPECT_EQ(ret2->LastModified, ret->LastModified);
-    EXPECT_EQ(ret2->SignedIdentifiers, identifiers);
+    auto ret2 = m_shareClient->GetAccessPolicy().ExtractValue();
+    EXPECT_EQ(ret2, identifiers);
   }
 
   TEST_F(FileShareClientTest, SharePermissions)
@@ -229,7 +225,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_FALSE(ret->FilePermissionKey.empty());
 
     auto ret2 = m_shareClient->GetPermission(ret->FilePermissionKey);
-    EXPECT_EQ(expectedPermission, ret2->FilePermission);
+    EXPECT_EQ(expectedPermission, *ret2);
   }
 
   // TEST_F(FileShareClientTest, Lease)
@@ -371,7 +367,7 @@ namespace Azure { namespace Storage { namespace Test {
     // Create/Get properties works
     std::unordered_map<std::string, Files::Shares::ShareClient> shareClients;
     std::string prefix = LowercaseRandomString(5);
-    Files::Shares::Models::GetSharePropertiesResult properties;
+    Files::Shares::Models::ShareProperties properties;
     {
       auto shareName = prefix + LowercaseRandomString(5);
       auto shareClient = Files::Shares::ShareClient::CreateFromConnectionString(
@@ -474,7 +470,7 @@ namespace Azure { namespace Storage { namespace Test {
     auto shareClient = Files::Shares::ShareClient::CreateFromConnectionString(
         PremiumFileConnectionString(), shareName);
     EXPECT_NO_THROW(shareClient.Create());
-    Files::Shares::Models::GetSharePropertiesResult properties;
+    Files::Shares::Models::ShareProperties properties;
     EXPECT_NO_THROW(properties = *shareClient.GetProperties());
     EXPECT_EQ(Files::Shares::Models::ShareAccessTier::Premium, properties.AccessTier.GetValue());
     EXPECT_FALSE(properties.AccessTierTransitionState.HasValue());